nouvelle version GSBparam

+++ controleurs/c_gestionProduits.php
@@ -41,10 +41,12 @@ if (isset($_SESSION['administrateur'])) {
                         );
 
                         echo 'Le Produit à bien été crée !';
-                    } else {
+                    }
+                    else {
                         echo 'Le produit existe déjà !';
                     }
-                } else {
+                }
+                else {
                     $contenances = getContenances();
                     $categories = getLesCategories();
                     $marques = getLesMarques();
@@ -57,14 +59,18 @@ if (isset($_SESSION['administrateur'])) {
         case 'creerContenance': {
 
                 if (isset($_POST['creer'])) {
-                    var_dump($_POST['laContenance']);
-                    var_dump($_POST['unite']);
+                    /*
+                     var_dump($_POST['laContenance']);
+                     var_dump($_POST['unite']);
+                     */
                     if (!existeContenance($_POST['laContenance'], $_POST['unite'])) {
                         creerContenance($_POST['laContenance'], $_POST['unite']);
-                    } else {
-                        echo 'Cette contenance exite déjà !';
                     }
-                } else {
+                    else {
+                        echo 'Cette contenance existe déjà !';
+                    }
+                }
+                else {
                     $unites = getLesUnites();
                     include("vues/v_creerContenance.php");
                 }
@@ -78,8 +84,16 @@ if (isset($_SESSION['administrateur'])) {
                     var_dump($_POST['contenance'], $_POST['id_contenance'], $_POST['stock'], $_POST['prix'], $_POST['idProd']);
                     updateStock($_POST['idProd'], $_POST['id_contenance'], $_POST['stock']);
                     updatePrix($_POST['idProd'], $_POST['id_contenance'], $_POST['prix']);
-                } else {
-                    supprimerPosseder($_POST['idProd'], $_POST['id_contenance']);
+                }
+                else {
+                    if (!verifContContenir($_POST['idProd'], $_POST['id_contenance'])) {
+                        supprimerPosseder($_POST['idProd'], $_POST['id_contenance']);
+                        echo 'Vous avez supprimé une contenance pour le produit ' . $_POST['idProd'] . ' !!';
+                    }
+                    else {
+                        echo 'Une commande posséde cette contenance, impossible de la supprimer !!';
+                    }
+
                 }
 
                 break;
@@ -95,7 +109,8 @@ if (isset($_SESSION['administrateur'])) {
                         $_POST['categorie'],
                     );
                     echo 'Le Produit à bien été modifié !';
-                } else {
+                }
+                else {
                     $contenances = getContenances();
                     $contenancesProd = getContenancesProd($_REQUEST['produit']);
                     //var_dump($contenancesProd);
@@ -115,10 +130,12 @@ if (isset($_SESSION['administrateur'])) {
                     if (!existeAssociation($_POST['produit1'], $_POST['produit2'])) {
                         associerProduit($_POST['produit1'], $_POST['produit2']);
                         echo 'Vous avez associer deux produits !';
-                    } else {
+                    }
+                    else {
                         echo 'Il y a déja une association entre ces deux produits !';
                     }
-                } else {
+                }
+                else {
                     if (isset($_POST['produit1']) && isset($_POST['produit2'])) {
                         if ($_POST['produit1'] == $_POST['produit2']) {
                             echo 'Deux mêmes produits !';
@@ -136,12 +153,23 @@ if (isset($_SESSION['administrateur'])) {
 
         case 'ajoutContenance': {
                 if (isset($_POST['submit'])) {
-                    var_dump($_REQUEST['produit']);
-                    var_dump($_REQUEST['lesContenances']);
-                    var_dump($_REQUEST['stock']);
-                    var_dump($_REQUEST['prix']);
-                    ajouterPosseder($_REQUEST['lesContenances'], $_REQUEST['produit'], $_REQUEST['stock'], $_REQUEST['prix']);
-                } else {
+                    /*
+                     var_dump($_REQUEST['produit']);
+                     var_dump($_REQUEST['lesContenances']);
+                     var_dump($_REQUEST['stock']);
+                     var_dump($_REQUEST['prix']);
+                     */
+                    if (!verifContPosseder($_REQUEST['produit'], $_REQUEST['lesContenances'])) {
+                        ajouterPosseder($_REQUEST['lesContenances'], $_REQUEST['produit'], $_REQUEST['stock'], $_REQUEST['prix']);
+                        echo 'Vous avez ajouté une nouvelle contenance à ' . $_REQUEST['produit'] . ' !!';
+                    }
+                    else {
+                        echo $_REQUEST['produit'] . ' possède déjà cette contenance !!';
+                    }
+
+
+                }
+                else {
                     $contenances = getContenances();
                     include("vues/v_ajoutContenance.php");
                 }
@@ -154,18 +182,21 @@ if (isset($_SESSION['administrateur'])) {
                 if (isset($_POST['modifier']) || isset($_POST['supprimer'])) {
                     if (isset($_POST['modifier'])) {
                         if (($_GET['produit1'] == $_POST['produit1'] || $_GET['produit1'] == $_POST['produit2']) &&
-                            ($_GET['produit2'] == $_POST['produit1'] || $_GET['produit2'] == $_POST['produit2'])
+                        ($_GET['produit2'] == $_POST['produit1'] || $_GET['produit2'] == $_POST['produit2'])
                         ) {
                             echo 'Modification impossible !';
-                        } else {
+                        }
+                        else {
                             if (!existeAssociation($_POST['produit1'], $_POST['produit2'])) {
                                 if ($_POST['produit1'] !== $_POST['produit2']) {
                                     modificationAssociation($_GET['produit1'], $_GET['produit2'], $_POST['produit1'], $_POST['produit2']);
                                     echo 'Vous avez modifier cette association !';
-                                } else {
+                                }
+                                else {
                                     echo 'Deux mêmes produits !';
                                 }
-                            } else {
+                            }
+                            else {
                                 echo 'Il y a déja une association entre ces deux produits !';
                             }
                         }
@@ -174,7 +205,8 @@ if (isset($_SESSION['administrateur'])) {
                         supprimerAssociation($_GET['produit1'], $_GET['produit2']);
                         echo 'Vous avez supprimer cette association !';
                     }
-                } else {
+                }
+                else {
                     $produits = getLesProduits();
                     include("vues/v_modifAssociation.php");
                 }
@@ -187,7 +219,8 @@ if (isset($_SESSION['administrateur'])) {
                 if (isset($_POST['modifStock'])) {
                     $lesStockProd = getStockProd($_POST['produitStock']);
                     include("vues/v_modifStock.php");
-                } else {
+                }
+                else {
                     $lesProduits = getLesProduits();
                     include("vues/v_produitStock.php");
                 }
@@ -203,6 +236,7 @@ if (isset($_SESSION['administrateur'])) {
                 //var_dump($_POST['stock']);
 
                 updateStock($_POST['idprod'], $_POST['idContenance'], $_POST['stock']);
+                echo 'Vous avez modifier le stock de ce produit ' . $_POST['idprod'] . ' !!';
 
                 break;
             }
@@ -213,8 +247,9 @@ if (isset($_SESSION['administrateur'])) {
                     supprimerAssociationsProd($_REQUEST['produit']);
                     supprimerProduit($_REQUEST['produit']);
                     echo 'Le produit à bien été supprimé';
-                } else {
-                    echo 'Il y a une commande avec ce produit, vous ne pouvez pas le supprimer !';
+                }
+                else {
+                    echo 'Une commande posséde ce produit, vous ne pouvez pas le supprimer !';
                 }
 
                 break;
@@ -228,7 +263,8 @@ if (isset($_SESSION['administrateur'])) {
                 break;
             }
     }
-} else {
+}
+else {
     switch ($action) {
         case 'connexion': {
                 //Si l'administrateur à déjà rentré les valeurs
@@ -239,26 +275,30 @@ if (isset($_SESSION['administrateur'])) {
                     if (count($msgErreurs) != 0) {
                         //S'il y a des erreurs alors on les affiches
                         include("vues/v_erreurs.php");
-                    } else {
+                    }
+                    else {
                         $connexion = connexionCompteAdministrateur($user, $password);
                         if (empty($connexion)) {
                             $msgErreurs[] = "Vous n'êtes pas inscrit.";
                             include("vues/v_erreurs.php");
-                        } else {
+                        }
+                        else {
                             $message = "Vous êtes désormais connecté";
                             include_once("vues/v_message.php");
                             $_SESSION['administrateur'] = $user;
                         }
                     }
-                } elseif (isset($_SESSION['administrateur']) and !empty($_SESSION['administrateur'])) {
+                }
+                elseif (isset($_SESSION['administrateur']) and !empty($_SESSION['administrateur'])) {
                     $message = "Vous est déjà authentifié";
                     include("vues/v_message.php");
-                } else {
+                }
+                else {
                     include("vues/v_connexionAdministrateur.php");
                 }
                 break;
             }
-            //Par défault l'on redirige l'utilisateur vers la page d'accueil pour cacher la page administrateur
+        //Par défault l'on redirige l'utilisateur vers la page d'accueil pour cacher la page administrateur
         default: {
                 header('Location:index.php?uc=accueil');
                 exit();
